Replace ProcessRunsTest
